+++ tests/test_model.py
@@ -87,7 +87,7 @@ class TestModel:
     def test_metadata_without_no_return(self):
         """Test metadata that doesn't have a return.
 
-        The function has an output and execution should ignore the output.
+        The function has an output, but execution should ignore the output.
         """
         G = Graph()
         G.add_node_object(Node("Test", lambda x: x))
@@ -124,19 +124,18 @@ class TestModel:
         )
 
     def test_model_visualize(self, model_instance):
-        """Test if the draw method of the model_instance.
+        """Test the visualize method of the model_instance.
 
-        The draw methods are tested in visualizer module. Here we make sure
-        the label is correct.
+        Here we make sure the label is correct.
         """
         dot_graph = model_instance.visualize()
 
         assert str(model_instance).replace("\n", r"\l") in dot_graph.source
 
-    def test_model_draw_export(self, model_instance, tmp_path):
-        """Test the draw method that exports to files.
+    def test_model_visualize_export(self, model_instance, tmp_path):
+        """Test the visualize method that exports to files.
 
-        Check the model description is in the file content.
+        Check that the model description is in the file content.
         """
 
         filename = tmp_path / "test_draw.dot"
@@ -324,7 +323,7 @@ class TestModelValidation:
     """Test is_graph_valid method of Model."""
 
     def test_is_valid_graph_digraph(self):
-        """Test is_graph_valid that correctly identifies non-directed graphs."""
+        """Test is_valid_graph that correctly identifies non-directed graphs."""
 
         G = nx.complete_graph(4)
         G.name = "test_graph"
@@ -335,16 +334,16 @@ class TestModelValidation:
             Model._is_valid_graph(G)
     
     def test_if_single_node_graph_is_valid(self, mmodel_G):
-        """Test is_graph_valid that correctly identifies single node graphs."""
+        """Test is_valid_graph that correctly identifies single-node graphs."""
         G = mmodel_G.__class__()
         G.add_node_object(Node("test", lambda x: x, output="c", inputs=["a"]))
 
         assert Model._is_valid_graph(G)
 
     def test_is_valid_graph_cycles(self):
-        """Test is_graph_valid that correctly identifies cycles.
+        """Test is_valid_graph that correctly identifies cycles.
 
-        Check the self-cycle and the nonself cycle.
+        Check both self-cycles and non-self cycles.
         """
 
         G = nx.DiGraph(name="test_graph")
@@ -366,7 +365,7 @@ class TestModelValidation:
             Model._is_valid_graph(G)
 
     def test_is_valid_graph_missing_attr(self, standard_G):
-        """Test is_graph_valid that correctly identifies isolated nodes.
+        """Test is_valid_graph that correctly identifies missing attributes.
 
         Here we add nodes to mmodel_G.
         """
@@ -445,6 +444,6 @@ class TestModelValidation:
             Model._is_valid_graph(G)
 
     def test_is_valid_graph_passing(self, mmodel_G):
-        """Test is_valid_graph that correctly passing."""
+        """Test is_valid_graph passing validation."""
 
         assert Model._is_valid_graph(mmodel_G)