+++ CHANGELOG.rst
@@ -28,9 +28,9 @@ Fixed
 Fixed
 ^^^^^^^
 
-- Fixe the issue that the "node_func" attribute can be overwritten during editing.
-- Fixe the issue that the "output" attribute of the edge is not updated during editing.
-- Fixe the issue of the different behavior of numpy.ufunc signature due to
+- Fix the issue that the "node_func" attribute can be overwritten during editing.
+- Fix the issue that the "output" attribute of the edge is not updated during editing.
+- Fix the issue of the different behavior of numpy.ufunc signature due to
   ``inspect.signature`` change in the Python version 3.11.9.
 - Rewrite node internal logic for converting function.